# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/tournaments": {
      "get": {
        "operationId": "TournamentsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Public Tournament page; every item includes derived favoriteCount and viewer-specific isFavorited.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/TournamentFavoriteViewFieldsDto"
                      }
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          }
        },
        "tags": [
          "Tournaments"
        ]
      },
      "post": {
        "operationId": "TournamentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTournamentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Tournaments"
        ]
      }
    },
...
    "/api/tournaments/{slug}/notifications": {
      "post": {
        "operationId": "NotificationController_createForTournament",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTournamentNotificationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Notification"
        ]
      }
    },
...
    "/api/notifications/{id}/read": {
      "patch": {
        "operationId": "NotificationController_markRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Notification"
        ]
      }
    },
    "/api/tournaments/{slug}/reports": {
      "post": {
        "operationId": "ReportController_create",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "reports"
        ]
      }
    },
    "/api/tournaments/{slug}/comments": {
      "post": {
        "operationId": "CommentController_create",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "comments"
        ]
      },
      "get": {
        "operationId": "CommentController_findByTournament",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "comments"
        ]
      }
    },
    "/api/comments/{id}/hide": {
      "patch": {
        "operationId": "CommentController_hide",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "comments"
        ]
      }
    },
    "/api/comments/{id}": {
      "delete": {
        "operationId": "CommentController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "comments"
        ]
      }
    },
...
    "/api/matches/{id}": {
      "get": {
        "operationId": "MatchesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Matches"
        ]
      },
      "patch": {
        "operationId": "MatchesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMatchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Matches"
        ]
      }
    },
    "/api/matches/bulk-schedule": {
      "patch": {
        "operationId": "MatchesController_bulkSchedule",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkScheduleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Matches"
        ]
      }
    },
    "/api/matches/{id}/scores": {
      "put": {
        "operationId": "MatchesController_putScores",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutMatchScoresDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Matches"
        ]
      }
    },
    "/api/rounds/{id}/matches": {
      "post": {
        "operationId": "MatchesController_createManual",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateManualMatchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Matches"
        ]
      }
    },
...
  },
  "info": {
    "title": "Esports Tournament Manager API",
    "description": "Backend API contract",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "TournamentFavoriteViewFieldsDto": {
        "type": "object",
        "properties": {
          "favoriteCount": {
            "type": "number",
            "example": 42,
            "minimum": 0
          },
          "isFavorited": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "favoriteCount",
          "isFavorited"
        ]
      },
...
      "CreateTournamentDto": {
        "type": "object",
        "properties": {
          "teamSize": {
            "type": "number",
            "description": "Số người thi đấu chính mỗi bên; backend kiểm tra theo metadata của Game",
            "minimum": 1,
            "maximum": 50
          },
          "customGameName": {
            "type": "string",
            "description": "Tên game do organizer đặt khi chọn canonical game CUSTOM",
            "maxLength": 100
          }
        }
      },
...
      "CreateTournamentNotificationDto": {
        "type": "object",
        "properties": {}
      },
      "CreateReportDto": {
        "type": "object",
        "properties": {}
      },
      "CreateCommentDto": {
        "type": "object",
        "properties": {}
      },
...
      "BulkScheduleDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateMatchDto": {
        "type": "object",
        "properties": {}
      },
      "PutMatchScoresDto": {
        "type": "object",
        "properties": {}
      },
      "CreateManualMatchDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}